requirements.txt
gradio>=4.29
typer>=0.12
toml>=0.10
pydantic>=2.6
watchdog>=4.0
psutil>=5.9
rich-click>=1.7
huggingface_hub>=0.23
boto3>=1.34
requests>=2.31
questionary>=2.0
rich>=13.7
sqlalchemy>=2.0
# PostgreSQL support
psycopg2-binary>=2.9
# Google Sheets integration
gspread>=6.1
google-auth>=2.29
# Dev / tooling
black>=24.4
ruff>=0.4
mypy>=1.10
pre-commit>=3.7
pytest>=8.2
torch>=2.2
transformers>=4.42
timm>=0.9
pillow>=10.3
plotly>=5.17
# ComfyUI integration
websocket-client>=1.6